Fixed an issue that an image will be created when string is empty
