+++ to.etc.domui/src/test/java/to/etc/domui/component/misc/TestTabInstance.java
@@ -1,0 +1,40 @@
+package to.etc.domui.component.misc;
+
+import org.junit.*;
+import to.etc.domui.component.layout.*;
+import to.etc.domui.dom.html.*;
+
+/**
+ * @author <a href="mailto:yoeri.nijs@itris.nl">Yoeri Nijs</a>
+ * Created on 18-8-17.
+ */
+public class TestTabInstance {
+
+	@Test
+	public void setImage_emptyString_expectNoImage() throws Exception {
+		// Setup
+		TabPanelBase tabPanelBase = new TabPanel(false);
+		tabPanelBase.tab();
+
+		// Execute
+		tabPanelBase.add(new Div("Some content"), "Content title", "");
+
+		// Verify
+		TabInstance tabInstance = tabPanelBase.getCurrentTabInstance();
+		Assert.assertNull(tabInstance.getImg());
+	}
+
+	@Test
+	public void setImage_noEmptyString_expectImage() throws Exception {
+		// Setup
+		TabPanelBase tabPanelBase = new TabPanel(false);
+		tabPanelBase.tab();
+
+		// Execute
+		tabPanelBase.add(new Div("Some content"), "Content title", "iconString");
+
+		// Verify
+		TabInstance tabInstance = tabPanelBase.getCurrentTabInstance();
+		Assert.assertNotNull(tabInstance.getImg());
+	}
+}
